requirements.txt
streamlit
serper
openrouter
python-dotenv
requests
pandas
